agent/requirements.txt
langgraph
python-dotenv
requests
mysql-connector-python